Handoff Push 2
More documentation and notetaking.
