+++ src/people/ct.js
@@ -6,7 +6,7 @@ import TakeTime from "../behavior/take-time.js";
 import task from "../support/task.js";
 import fluentBehaviorTree from "@crowdedjs/fluent-behavior-tree"
 
-class ctThesis {
+class ct {
 
     constructor(myIndex) {
         this.index = myIndex;
@@ -38,8 +38,7 @@ class ctThesis {
         
         this.tree = builder
 
-        // Consider limiting the rooms nurses can be assigned to tasks to
-        // General Structure of New Trees: GO TO START -> GET A TASK -> GO TO THE TASK -> ACCOMPLISH THE TASK FROM *LIST OF TASKS* AND TAKE TIME -> RESTART
+        // STRUCTURE OF TREES: TESTING -> GO TO START -> QUEUE STORED TASKS -> GET A TASK -> GO TO THE TASK -> ACCOMPLISH THE TASK FROM *LIST OF TASKS* AND TAKE TIME -> RESTART
         .parallel("Testing Parallel", 2, 2)
             .do("Testing", (t) => {
                 if (me().onTheClock && me().getTask() == null && me().taskTime == 0 && !me().moving) {
@@ -231,4 +230,4 @@ class ctThesis {
     }
 }
 
-export default ctThesis;
+export default ct;
