+++ src/people/old_trees/agent.js
@@ -1,3 +1,4 @@
+// BASE AGENT CLASS, NOT USED ANYMORE BUT IS TEMPLATE FOR PATIENT-AGENT AND MEDICAL-AGENT
 class Agent {
   startX;
   startY;
