+++ src/people/old_trees/escape-person.js
@@ -1,4 +1,4 @@
-import GoTo from "../behavior/go-to.js"
+import GoTo from "../../behavior/go-to.js"
 import fluentBehaviorTree from "@crowdedjs/fluent-behavior-tree"
