+++ src/people/old_trees/get-patient.js
@@ -1,4 +1,4 @@
-import GoTo from "../behavior/go-to.js";
+import GoTo from "../../behavior/go-to.js";
 import fluentBehaviorTree from "@crowdedjs/fluent-behavior-tree"
