+++ src/people/tech.js
@@ -7,7 +7,7 @@ import task from "../support/task.js";
 import PatientTempState from "../support/patient-temp-state.js";
 import fluentBehaviorTree from "@crowdedjs/fluent-behavior-tree"
 
-class techThesis {
+class tech {
 
     constructor(myIndex) {
         this.index = myIndex;
@@ -27,6 +27,8 @@ class techThesis {
 
         this.tree = builder
 
+        // Consider limiting the rooms nurses can be assigned to tasks to
+        // STRUCTURE OF TREES: TESTING -> GO TO START -> QUEUE STORED TASKS -> GET A TASK -> GO TO THE TASK -> ACCOMPLISH THE TASK FROM *LIST OF TASKS* AND TAKE TIME -> RESTART
         .parallel("Testing Parallel", 2, 2)
             .do("Testing", (t) => {
                 // This would tick up while on the way back to the computer, which isn't desirable.
@@ -44,8 +46,6 @@ class techThesis {
                 me().lengthOfStay++;
                 return fluentBehaviorTree.BehaviorTreeStatus.Running; 
             })
-        // Consider limiting the rooms nurses can be assigned to tasks to
-        // General Structure of New Trees: GO TO START -> GET A TASK -> GO TO THE TASK -> ACCOMPLISH TASK FROM LIST -> TAKE TIME -> QUEUE TASKS -> RESTART
         .sequence("Tech Behaviors")
             
             .do("Testing", (t) => {
@@ -303,4 +303,4 @@ class techThesis {
     }
 }
 
-export default techThesis;
+export default tech;
