+++ src/people/xray.js
@@ -6,7 +6,7 @@ import TakeTime from "../behavior/take-time.js";
 import task from "../support/task.js";
 import fluentBehaviorTree from "@crowdedjs/fluent-behavior-tree"
 
-class radiologyThesis {
+class xray {
 
     constructor(myIndex) {
         this.index = myIndex;
@@ -17,29 +17,36 @@ class radiologyThesis {
         let self = this;//Since we need to reference this in anonymous functions, we need a reference
         let me= ()=>Hospital.agents.find(a=>a.id == myIndex);
         
-        let goToName = "CT 2";
-        let myGoal = Hospital.locations.find(l => l.name == goToName);
-        let computer = Hospital.locations.find(l => l.name == goToName);
-        let entrance = Hospital.getLocationByName("Main Entrance");
-
-        if (!myGoal) {
-            myGoal = Hospital.locations.find(l => l.name == "CT 1");
+        let goToName = "XRay 1";
+        if (myIndex % 2 == 1) {
+            goToName = "XRay 2";
         }
 
-        let taskQueue = [];
+        let myGoal = Hospital.locations.find(l => l.name == goToName);
+        let computer = Hospital.locations.find(l => l.name == goToName);
+        let xray1 = Hospital.locations.find(l => l.name == "XRay 1");
+        let xray2 = Hospital.locations.find(l => l.name == "XRay 2");
+        //if (!myGoal) throw new exception("We couldn't find a location called " + goToName);
+        if (!myGoal) { 
+            myGoal = xray1;
+            computer = xray1;
+        };
+        let entrance = Hospital.getLocationByName("Main Entrance");
 
+        let taskQueue = [];
+        
         this.tree = builder
 
+        // STRUCTURE OF TREES: TESTING -> GO TO START -> QUEUE STORED TASKS -> GET A TASK -> GO TO THE TASK -> ACCOMPLISH THE TASK FROM *LIST OF TASKS* AND TAKE TIME -> RESTART
         .parallel("Testing Parallel", 2, 2)
             .do("Testing", (t) => {
-                // This would tick up while on the way back to the computer, which isn't desirable.
                 if (me().onTheClock && me().getTask() == null && me().taskTime == 0 && !me().moving) {
                     me().idleTime++;
                 }
                 if (me().lengthOfStay == 43200 || me().lengthOfStay == 86399) {
                     let idleTimeMinutes = ((1440 * me().idleTime) / 86400);
                     idleTimeMinutes = Math.round((idleTimeMinutes + Number.EPSILON) * 100) / 100
-                    //console.log("Radiology Idle Time: " + me().idleTime + " ticks / " + idleTimeMinutes + " minutes in-simulation");
+                    //console.log("X-Ray Idle Time: " + me().idleTime + " ticks / " + idleTimeMinutes + " minutes in-simulation");
                     console.log(idleTimeMinutes);
                     me().idleTime = 0;
                     //me().lengthOfStay = 0;
@@ -47,31 +54,34 @@ class radiologyThesis {
                 me().lengthOfStay++;
                 return fluentBehaviorTree.BehaviorTreeStatus.Running; 
             })
-        // General Structure of New Trees: GO TO START -> GET A TASK -> GO TO THE TASK -> ACCOMPLISH THE TASK FROM *LIST OF TASKS* AND TAKE TIME -> RESTART
-        .sequence("Radiology Behaviors")
+        .sequence("XRay Behaviors")
             
             .do("Testing", (t) => {
                 me().moving = true;
                 return fluentBehaviorTree.BehaviorTreeStatus.Success;            
             })
-            
+        
             .splice(new GoTo(self.index, myGoal.location).tree)
-            //.splice(new AssignBed(myIndex, Hospital.locations.find(l => l.name == "C1").location).tree)
-            
+
             .do("Testing", (t) => {
                 me().moving = false;
                 return fluentBehaviorTree.BehaviorTreeStatus.Success;            
             })
 
-            .splice(new AssignComputer(myIndex, myGoal.location).tree) // CT 2
-            
-            // Add a behavior here or in the selector that will order the tasks (by severity)?
-            
+            .splice(new AssignComputer(myIndex, computer.location).tree) // RESIDENT PLACE
+
+            // QUEUEING FOLLOWING TASKS NEEDS TO COME LAST, OTHERWISE TASKS ARE BLITZED THROUGH TOO QUICKLY
             .do("Queue Tasks", (t) => {
                 while (taskQueue.length > 0) {
                     switch(taskQueue[0].taskID) {
-                        case "Resident Scan Read":
-                            Hospital.residentTaskList.push(taskQueue.shift());
+                        case "Pick Up Patient":
+                            Hospital.techTaskList.push(taskQueue.shift());
+                            break;
+                        case "XRay Pickup":
+                            Hospital.techTaskList.push(taskQueue.shift());
+                            break;
+                        case "Radiology Review Scan":
+                            Hospital.radiologyTaskList.push(taskQueue.shift());
                             break;
                         default: break;
                     }
@@ -80,17 +90,48 @@ class radiologyThesis {
                 return fluentBehaviorTree.BehaviorTreeStatus.Success;
             })
 
+            // Add a behavior here or in the selector that will order the tasks (by severity)?
             .selector("Task List Tasks")
+                // THIS TASK IS GIVEN BY THE XRAY TO THE TECH AND MUST BE DONE BEFORE GETTING A TASK
+                // THE XRAY QUEUE IS FILLED BY THE RESIDENT
+                .do("Queue Escort Patient", (t) => {
+                    if (Hospital.XRayQueue.length == 0 || (goToName == "XRay 1" && Hospital.isXRay1Occupied()) || (goToName == "XRay 2" && Hospital.isXRay2Occupied())) {
+                        return fluentBehaviorTree.BehaviorTreeStatus.Failure;
+                    }
+                    else {
+                        // Pick Up Patient
+                        let xrayPatient = Hospital.XRayQueue.shift();
+                        let techEscortTask = new task("Pick Up Patient", xrayPatient.getSeverity(), 0, xrayPatient, myGoal);
+                        if (goToName == "XRay 1") {
+                            Hospital.setXRay1Occupied(true);
+                            xrayPatient.setImagingRoom("XRay 1");
+                            techEscortTask = new task("Pick Up Patient", xrayPatient.getSeverity(), 0, xrayPatient, xray1);
+                        }
+                        else if (goToName == "XRay 2") {
+                            Hospital.setXRay2Occupied(true);
+                            xrayPatient.setImagingRoom("XRay 2");
+                            techEscortTask = new task("Pick Up Patient", xrayPatient.getSeverity(), 0, xrayPatient, xray2);
+                        }
+                        //Hospital.techTaskList.push(techEscortTask);
+                        taskQueue.push(techEscortTask);
+                    }
+
+                    return fluentBehaviorTree.BehaviorTreeStatus.Success;
+                })
+
                 .do("Get a Task", (t) => {
                     // CHECK IF NEEDED TO CLOCK IN
                     if (!me().onTheClock) {
                         me().onTheClock = true;
-                        Hospital.activeRadio.push(me());
-                        
+                        Hospital.activeXRay.push(me());
+                        if (Hospital.aTeam[5] == null) {
+                            Hospital.aTeam[5] = me();
+                        }
+
                         return fluentBehaviorTree.BehaviorTreeStatus.Success;
                     }
                     // CHECK IF NEEDED TO CLOCK OUT
-                    else if (Hospital.activeRadio.length > 1 && Hospital.activeRadio[0] == me()) {
+                    else if (Hospital.activeXRay.length > 2 && Hospital.activeXRay[0] == me()) {
                         let clockOutTask = new task("Clock Out", null, null, null, entrance);
                         me().setTask(clockOutTask);
                         me().replacement = true;
@@ -101,8 +142,8 @@ class radiologyThesis {
                         return fluentBehaviorTree.BehaviorTreeStatus.Failure;
                     }
                     // CHECK IF ANY TASKS ARE AVAILABLE, CONTINUE
-                    else if (Hospital.radiologyTaskList.length != 0) {
-                        me().setTask(Hospital.radiologyTaskList.shift());
+                    else if (Hospital.xrayTaskList.length != 0) {
+                        me().setTask(Hospital.xrayTaskList.shift());
                         return fluentBehaviorTree.BehaviorTreeStatus.Failure;
                     }
                     // OTHERWISE DON'T PROCEED (SUCCESS WILL RESTART SELECTOR)
@@ -110,7 +151,7 @@ class radiologyThesis {
                         return fluentBehaviorTree.BehaviorTreeStatus.Success;
                     }
                 })
-                // I think it has to be done this way because you can't do operations in a splice for the most part
+                
                 .inverter("Need to return failure")
                     .sequence("Go to Task")
                         .do("Determine Location", (t) => {
@@ -118,6 +159,7 @@ class radiologyThesis {
                                 myGoal = me().getTask().location;
                             }
                             else {
+                                //myGoal = Hospital.locations.find(l => l.name == goToName);
                                 myGoal = computer;
                             }
                             return fluentBehaviorTree.BehaviorTreeStatus.Success; 
@@ -132,36 +174,51 @@ class radiologyThesis {
                         return fluentBehaviorTree.BehaviorTreeStatus.Failure;
                     }
                     else {
-                        Hospital.activeRadio.shift();
-                        
-                        // TESTING
-                        let idleTimeMinutes = ((1440 * me().idleTime) / 86400);
-                        idleTimeMinutes = Math.round((idleTimeMinutes + Number.EPSILON) * 100) / 100
-                        console.log("Radiology Idle Time: " + me().idleTime + " ticks / " + idleTimeMinutes + " minutes in-simulation");
-                        Hospital.radioData.push(me().idleTime);
-
+                        // for(let i = 0; i < Hospital.computer.entries.length; i++) {
+                        //     if (Hospital.computer.entries[i].getTech() == me()) {
+                        //       Hospital.computer.entries[i].setTech(null);
+                        //     }
+                        // }
+                        if (Hospital.aTeam[5] == me()) {
+                            Hospital.aTeam[5] = null;
+                        }
+                        Hospital.activeXRay.shift();
                         me().inSimulation = false;
                         return fluentBehaviorTree.BehaviorTreeStatus.Running;
                     }
                 })
                 
-                // THIS TASK IS GIVEN BY THE CT (CAT Do Scan Behavior)
-                .do("Radiology Review Scan", (t) => {
-                    if (me().getTask().taskID != "Radiology Review Scan") {
+                //THIS TASK IS GIVEN BY THE TECH  
+                .do("XRay Do Scan", (t) => {
+                    if (me().getTask().taskID != "XRay Do Scan") {
                         return fluentBehaviorTree.BehaviorTreeStatus.Failure;
                     }
                     else {
+                        // 30 minutes = 1800 ticks
+                        //me().taskTime = 1800;
                         me().taskTime = 60;
-                        
-                        let residentScanTask = new task("Resident Scan Read", null, null, me().getTask().patient, null);
-                        taskQueue.push(residentScanTask);
-                        //Hospital.residentTaskList.push(residentScanTask);
+                        me().getTask().patient.setScan(true);
+                        //me().getTask().patient.waitInScanRoom = true;
+                        let xrayPickupTask = new task("XRay Pickup", null, 0, me().getTask().patient, myGoal);
+                        if (me().getTask().patient.getImagingRoom() == "XRay 1") {
+                            xrayPickupTask = new task("XRay Pickup", null, 0, me().getTask().patient, xray1);
 
+                        }
+                        else {
+                            xrayPickupTask = new task("XRay Pickup", null, 0, me().getTask().patient, xray2);
+                        }
+                        taskQueue.push(xrayPickupTask);
+                        //Hospital.techTaskList.push(xrayPickupTask);
+
+                        let radiologyReviewTask = new task("Radiology Review Scan", null, 0, me().getTask().patient, null);
+                        taskQueue.push(radiologyReviewTask);
+                        //Hospital.radiologyTaskList.push(radiologyReviewTask);
+                        
                         me().setTask(null);
                         return fluentBehaviorTree.BehaviorTreeStatus.Success;
                     }
-                })
-          
+                }) 
+                
             .end()
             // IF SUCCEEDING IN TASK, TAKE TIME TO DO THAT TASK
             // TakeTime doesn't work in some instances, but the code itself works. For instance if you remove the next .end(), it will work, but then the sequence is broken.
@@ -173,7 +230,7 @@ class radiologyThesis {
                     return fluentBehaviorTree.BehaviorTreeStatus.Running;
                 }
                 return fluentBehaviorTree.BehaviorTreeStatus.Success;
-            })            
+            })
         .end()
         .end()
         .build()
@@ -184,4 +241,4 @@ class radiologyThesis {
     }
 }
 
-export default radiologyThesis;
+export default xray;
