+++ src/support/acomment.js
@@ -1,3 +1,5 @@
+// PORTED CODE FROM JAVA IMPLEMENTATION
+// WOULD BE USED FOR CREATING COMMENT DIALOGUE BUBBLES ABOVE AGENTS, NEVER GOT REIMPLEMENTED
 class AComment {
 
 	get Text() {
